Send 403 if contact does nor belong to user

+++ delete.php
@@ -20,6 +20,14 @@
      return;
    }
 
+   $contact = $statement->fetch(PDO::FETCH_ASSOC);
+
+   if ($contact["user_id"] !== $_SESSION["user"]["id"]) {
+     http_response_code(403);
+     echo("HTTP 403 UNAUTORIZED");
+     return;
+   }
+
    $conn->prepare("DELETE FROM contacts WHERE id = :id")->execute([":id" => $id]);
 
    header("Location: home.php");
